Updated so that all visual properties are now supported.
git-svn-id: http://chianti.ucsd.edu/svn@16768 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ core3/ding-presentation-impl/trunk/src/main/java/org/cytoscape/ding/PrintLOD.java
@@ -52,7 +52,7 @@ public class PrintLOD extends GraphLOD {
 	 *
 	 * @return  DOCUMENT ME!
 	 */
-	public byte renderEdges(int visibleNodeCount, int totalNodeCount, int totalEdgeCount) {
+	public int renderEdges(int visibleNodeCount, int totalNodeCount, int totalEdgeCount) {
 		return 0;
 	}
